# docs/swagger.yaml
definitions:
  models.Article:
    properties:
      author_id:
        type: string
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - author_id
    type: object
  models.Author:
    properties:
      created_at:
        type: string
      fullname:
        example: John Doe Steve
        maxLength: 255
        minLength: 2
        type: string
      id:
        type: string
      updated_at:
        type: string
    required:
    - fullname
    type: object
  models.CreateArticleModel:
    properties:
      author_id:
        type: string
      body:
        type: string
      title:
        type: string
    required:
    - author_id
    type: object
  models.CreateAuthorModel:
    properties:
      fullname:
        example: John Doe Steve
        maxLength: 255
        minLength: 2
        type: string
    required:
    - fullname
    type: object
  models.DeleteArticleModel:
    properties:
      author:
        $ref: '#/definitions/models.Author'
      body:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.JSONErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.JSONResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  models.PackedArticleModel:
    properties:
      author:
        $ref: '#/definitions/models.Author'
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.UpdateArticleModel:
    properties:
      body:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  models.UpdateAuthorModel:
    properties:
      fullname:
        example: John Doe Steve
        maxLength: 255
        minLength: 2
        type: string
      id:
        type: string
    required:
    - fullname
    type: object
info:
  contact:
    email: john.doe@swagger.io
    name: API Article
    url: https://john.doe.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /v2/article:
    get:
      consumes:
      - application/json
      description: get articles
      parameters:
      - description: "0"
        in: query
        name: offset
        type: integer
      - description: "10"
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Article'
                  type: array
              type: object
      summary: List articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: create a new article
      parameters:
      - description: article body
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/models.CreateArticleModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Create article
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: update a new article
      parameters:
      - description: article body
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/models.UpdateArticleModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Article'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: update article
      tags:
      - articles
  /v2/article/{id}:
    delete:
      consumes:
      - application/json
      description: delete an article by id
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.DeleteArticleModel'
              type: object
        "404":
          description: Not Found
          schema: {}
      summary: delete article by id
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: get an article by id
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PackedArticleModel'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: get article by id
      tags:
      - articles
  /v2/author:
    get:
      consumes:
      - application/json
      description: get author
      parameters:
      - description: "0"
        in: query
        name: offset
        type: integer
      - description: "10"
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Author'
                  type: array
              type: object
      summary: List author
      tags:
      - authors
    post:
      consumes:
      - application/json
      description: create a new author
      parameters:
      - description: author body
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/models.CreateAuthorModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Author'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Create author
      tags:
      - authors
    put:
      consumes:
      - application/json
      description: update a new author
      parameters:
      - description: author body
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAuthorModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Author'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: update author
      tags:
      - authors
  /v2/author/{id}:
    delete:
      consumes:
      - application/json
      description: delete an author by id
      parameters:
      - description: author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Author'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: delete author by id
      tags:
      - authors
    get:
      consumes:
      - application/json
      description: get an author by id
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Author'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: get author by id
      tags:
      - authors
swagger: "2.0"